Methods findByIds and saveAll from now on do not depend on the name of the id field.

+++ src/main/java/pg/hib/entities/TestEntity.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 public class TestEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column
     private boolean active;
@@ -24,17 +24,17 @@ public class TestEntity implements Serializable {
         this(null, active, created);
     }
 
-    public TestEntity(Long id, boolean active, LocalDateTime created) {
+    public TestEntity(Integer id, boolean active, LocalDateTime created) {
         this.id = id;
         this.active = active;
         this.created = created;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
